aws/backend/requirements.txt
flask

pymongo
dnspython

python-dotenv
gunicorn
flask-cors